+++ src/main/java/calculatorUtils/FactorialUtils.java
@@ -5,12 +5,12 @@ import org.apache.logging.log4j.Logger;
 
 public class FactorialUtils {
     public static final Logger log = LogManager.getLogger(FactorialUtils.class);
-    public Integer computeFactorial(Integer number){
-        Integer factorial = 1;
+    public Long computeFactorial(Integer number){
+        Long factorial = 1l;
         if(number < 0) {
-            factorial = -1;
+            factorial = -1l;
         }else if(number == 0 || number == 1) {
-            factorial = 1;
+            factorial = 1l;
         }else {
             for (int count = number; count > 1; count--)
                 factorial *= count;
